fix: use pipeline ns when generating binding names
- relates to issue #266
